ai-engine/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# ML and Chess
torch==2.1.2
onnxruntime==1.16.3
numpy==1.26.3
python-chess==1.999

# Training
tensorboard==2.15.1
mlflow==2.10.0

# Utilities
requests==2.31.0
aiohttp==3.9.1
redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
